Move parse_source/1 to Parser
